(#112) Correçaõ e retirada de empilhamento em relatorio.js
