+++ src/modules/apiServices.js
@@ -1,4 +1,5 @@
-import {GraficoPessoal} from '../Chart/app.js';
+import {GraficoPessoal} from '../libraries/app.js';
+
 export function CodigoDevicePost(){
     var xhr = new XMLHttpRequest();
 
@@ -162,9 +163,9 @@ export function contribuinteRepositorio(numeroMilestone, token, $owner, $repo, s
    
    
    if(contador == numerosContribuintes - 1){
-     console.log(nomeContribuinte)
-     console.log(qtdComitsContribuinte)
-     GraficoPessoal(nomeContribuinte, qtdComitsContribuinte, nomeSprint)
+     console.log(nomeContribuinte);
+     console.log(qtdComitsContribuinte);
+     GraficoPessoal(nomeContribuinte, qtdComitsContribuinte, nomeSprint);
    }
     
    contador++;
