+++ src/popup/popup.html
@@ -22,7 +22,7 @@
 
     <script type='text/javascript' src="pdfmakeMin.js"></script>
     <script type='text/javascript' src="pdfmakeFont.js"></script>
-    <script type='text/javascript' src='../Chart/chart.js'></script>
+    <script type='text/javascript' src='../libraries/chart.js'></script>
 
     <script type='module' src='popup.js'></script>
